+++ src/test/java/tudelft/ti2806/pl3/util/EmptyEdgesFourNodeSpace.java
@@ -1,4 +1,4 @@
-package tudelft.ti2806.pl3.visualisation;
+package tudelft.ti2806.pl3.util;
 
 import org.junit.Before;
 import org.junit.Test;
